refactor: move AdminFormPage into features/admin-form/common
clean up hierarchy
