+++ frontend/src/features/admin-form/common/components/AdminFormNavbar/AdminFormNavbarDetails.tsx
